Fix drawing scaled ellipses with border
